@@ -2,7 +2,7 @@ module DevisePermittedParameters
   extend ActiveSupport::Concern
 
   included do
-    before_filter :configure_permitted_parameters
+    before_action :configure_permitted_parameters
   end
 
   protected
